autogen/titanic_analysis_team/requirements.txt
python-dotenv
matplotlib
seaborn
pandas
numpy
pydantic
pyautogen
autogen-agentchat
autogen-ext[openai]
langchain_core
rich
wikipedia-mcp
fastmcp
wikipedia-api
requests
httpx
pydantic[email]
seaborn